Implement GPU acceleration for the hydro treewalk
